requirements.txt
streamlit==1.41.1
Pillow==10.4.0
